Move info.Symbol under c_symbols.
